backend/requirements.txt
Django==5.1.2
django-cors-headers==4.6.0
numpy==2.2.4
packaging==24.2
Pygments==2.18.0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
djangorestframework==3.16.0
